# Conflicts:
#	src/main/java/com/mbicycle/imarket/beans/dto/OrderDTO.java
#	src/main/java/com/mbicycle/imarket/facades/impl/OrderFacadeImpl.java
#	src/main/java/com/mbicycle/imarket/services/impl/OrderServiceImpl.java
#	src/main/java/com/mbicycle/imarket/utils/converters/impl/OrderConverter.java
#	src/main/java/com/mbicycle/imarket/utils/converters/impl/reversed/ReverseOrderConverter.java

+++ src/main/java/com/mbicycle/imarket/services/impl/OrderServiceImpl.java
@@ -15,13 +15,10 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public boolean add(Order order) {
-        orderRepository.save(order);
-        return true;
-    }
-
-    @Override
-    public boolean update(Order order) {
-        return false;
+        if (findByProfile(order.getProfile()) == null) {
+            orderRepository.save(order);
+        }
+        return findByProfile(order.getProfile()) != null;
     }
 
     @Override
@@ -35,8 +32,30 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
+    public boolean update(Order orderToUpdate) {
+        Order order = orderRepository.findByProfile(orderToUpdate.getProfile());
+        if (order != null) {
+            order.setDateClosed(orderToUpdate.getDateClosed());
+            order.setDateGot(orderToUpdate.getDateGot());
+            order.setDateOpened(orderToUpdate.getDateOpened());
+            order.setDatePaid(orderToUpdate.getDatePaid());
+            order.setDateReady(orderToUpdate.getDateReady());
+            order.setDateSent(orderToUpdate.getDateSent());
+            orderRepository.save(order);
+        }
+        return order.equals(orderToUpdate);
+    }
+
+    @Override
+    public Order findByProfile(Profile profile) {
+        return orderRepository.findByProfile(profile);
+    }
+
+    @Override
     public boolean delete(Order order) {
-        orderRepository.delete(order);
-        return true;
+        if (findByProfile(order.getProfile()) != null) {
+            orderRepository.delete(order);
+        }
+        return findByProfile(order.getProfile()) == null;
     }
 }
